requirements.txt
streamlit>=1.32.0
groq>=0.4.2
python-dotenv>=1.0.0
PyPDF2>=3.0.0
sentence-transformers>=3.4.1
scikit-learn>=1.3.0
numpy>=1.26.2
tqdm>=4.66.1
watchdog>=4.0.0